web-analyst-agent/requirements.txt
streamlit==1.31.0
openai==1.12.0
faiss-cpu==1.7.4
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.3
readability-lxml==0.8.1
youtube-transcript-api==0.6.2
python-dotenv==1.0.1
tiktoken==0.6.0
pydantic==2.6.0
pytest==8.0.0
